+++ pom.xml
@@ -25,6 +25,7 @@
 	</properties>
 	<modules>
 		<module>spring-cloud-stream</module>
+		<module>spring-cloud-stream-rxjava</module>
 		<module>spring-cloud-stream-binders</module>
 		<module>spring-cloud-stream-codec</module>
 		<module>spring-cloud-stream-module-launcher</module>
