+++ spring-cloud-stream-rxjava/src/main/java/org/springframework/cloud/stream/annotation/rxjava/EnableRxJavaModule.java
@@ -23,14 +23,9 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.stream.annotation.EnableModule;
 import org.springframework.cloud.stream.annotation.Processor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.integration.annotation.ServiceActivator;
-import org.springframework.messaging.MessageHandler;
 
 /**
  * Annotation that identifies the class as RxJava module. The class that has {@link EnableRxJavaModule} annotated is expected
@@ -43,21 +38,6 @@ import org.springframework.messaging.MessageHandler;
 @Documented
 @Inherited
 @EnableModule(Processor.class)
-@Import(EnableRxJavaModule.RxJavaMessageHandler.class)
+@Import(RxJavaConfiguration.class)
 public @interface EnableRxJavaModule {
-
-	@Configuration
-	class RxJavaMessageHandler {
-
-		@Autowired
-		RxJavaProcessor processor;
-
-		@ServiceActivator(inputChannel = Processor.INPUT, outputChannel = Processor.OUTPUT)
-		@Bean
-		public MessageHandler subjectMessageHandler() {
-			SubjectMessageHandler messageHandler = new SubjectMessageHandler(processor);
-			messageHandler.setOutputChannelName(Processor.OUTPUT);
-			return messageHandler;
-		}
-	}
 }
